PIM Manager: implement Manager.Stop() properly and fix delayed starting
Instead of always stopping, first check that there are no views in use
by a client. This can be determined based on the use count of the full
view.

For testing this, an internal, undocumented Manager.IsRunning() D-Bus
method gets added.

While writing the corresponding test, it became obvious that setting the sort
order had the undesired side effect of starting up folks. Aggregation still
didn't run (that gets started by ViewResource), but even just having folks
running is too much for a syncevo-dbus-server that is just used for syncing.