D-Bus server: enhance asynchronous result reporting
Grant access to stored Result callbacks, needed because PIM manager
needs to report failures and can do that more efficiently when keeping
a copy of just the error callback, instead of adding yet another
boost::bind on top of Result.

Also add factory methods which wrap the D-Bus result object in a
Result instance of the right type. Also needed for PIM manager.