GDBus GIO: add support for deque and list
The actual implementation is the same as for vector, with D-Bus array
on the D-Bus side. Now the common code is in a template base class and
referenced by the traits for vector, list and deque.

No attempt is made to automatically pick that trait for other classes
which also could be treated like an ordered collection of items.