PIM: add --debug to sync.py example
The --debug command line flag enables printing of the debug output
sent by SyncEvolution over D-Bus. To capture it in a timely manner,
the running method calls must be done asynchronously with a running
main loop.

This also solves the timeout problem more nicely and is closer to how
a real application should be written.