Folks: fix refining a search
Due to not mapping the local index in the view to the parent's index,
refining only worked in views where parent and child had the same
index for the contacts in the search view.

Make the tests more complex to trigger this bug and fix it by adding
the required indirection.