PIM: refactor folks classes
Move IndividualView, FullView and FilteredView into separate source
files. They are large enough to justify that and it'll avoid header
file dependency cycles.

Introduce a new View base class, which contains the "start"
functionality and a new name for debug messages.

The new StreamingView is based on it and will be used for reporting
results in an EDS query.

Avoid passing NULL IndividualCompare pointer to views. Instead make
the CompareFormattedName class available as the default compare
method.