syncevo-dbus-server: log only to syslog by default
The new default is to log error messages to syslog.
Command line options can increase (or reduce) verbosity
and choose whether stdout and/or syslog shall be active:

  -d, --duration=seconds/'unlimited'
      Shut down automatically when idle for this duration
  -v, --verbosity=level
      Choose amount of output, 0 = no output, 1 = errors,
      2 = info, 3 = debug; default is 1.
  -o, --stdout
      Enable printing to stdout (result of operations) and
      stderr (errors/info/debug).
  -s, --no-syslog
      Disable printing to syslog.

To get the same behavior as before when debugging via the Python
scripts, "--no-syslog --stdout --verbosity=3" is passed to
syncevo-dbus-server when started via Python.