PIM: fix memory leaks during writing of contacts
Constructing the GValues created additional references instead
of taking over ownership as intended. For refcounted GObject,
the solution is to pass the instance and let GValue take the
reference.

For other cases, ownership of a new instance is transfered to the
GValue using the new constructors with a boolean flag.