D-Bus testing: improved logging
Log message now contain time stamps. A NullLogging class mimics the
Logging interface and can be used instead of that to suppress logging.

As a side effect of turning the log() method into a wrapper, the D-Bus
signal is now called "log2", which makes it possible to search for it
case-insensitively in emacs without finding the LogOutput signal.