LogRedirect: better handling of realloc failure
When realloc() returns NULL, the already allocated buffer remains
available. Previously we discarded (and leaked!) it. Now we track
whether that memory already contains parts of the next log message
and use that when realloc fails. The dgram will then get discarded
without reading the rest.

When reading from a stream, realloc will only be called once to
get the initial buffer. When that fails, we give up.