D-Bus: fix random session failures
When the helper shuts down normally after the parent is done with it,
there was a race between handling the signal and the loss of connection,
leading to random "return 1" errors. This showed up in nightly testing
in particular in the test-dbus.py testConfigure.

The loss of connection at that point is not an error, so don't treat it
as one and simply return 0.